quiz/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
classopt==0.2.1
click==8.1.3
cmake==3.26.0
colorama==0.4.6
comm==0.1.2
compare-mt==0.2.10
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
filelock==3.10.0
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
fugashi==1.2.1
ginza==5.1.2
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
huggingface-hub==0.13.2
idna==3.4
ipadic==1.0.0
ipykernel==6.21.3
ipython==8.11.0
ja-ginza==5.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.3.0
kiwisolver==1.4.4
langcodes==3.3.0
lightning-utilities==0.8.0
lit==15.0.7
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mecab-python3==1.0.5
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
neologdn==0.5.1
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
plac==1.3.5
platformdirs==3.1.1
portalocker==2.7.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==2.0.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
smart-open==6.3.0
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
SudachiDict-core==20230110
SudachiPy==0.6.7
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thinc==8.1.9
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==2.0.0
torchaudio==2.0.1
torchdata==0.6.0
torchmetrics==0.11.4
torchtext==0.15.1
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.1
triton==2.0.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==0.10.1
wcwidth==0.2.6
Werkzeug==2.2.3
yarl==1.8.2
